Rate limit, daily, spelling
Added a rate limit to the backend re did the frontend for the daily page. And re did some spelling making files camel case.
