+++ FitnessApp/src/Routing/indexRoute.jsx
@@ -3,9 +3,9 @@ import PrivateRoutes from "./privateRoutes.jsx";
 import PublicRoutes from "./publicRoutes.jsx";
 import StartPage from "../pages/Landing/StartPage.jsx";
 import LoginPage from "../pages/login/login.jsx";
-import Daily from "../pages/Daily/Daily.jsx";
+import Daily from "../pages/daily/daily.jsx";
 import Share from "../pages/share/share.jsx";
-import ViewPage from "../pages/DataView/ViewPage.jsx";
+import ViewPage from "../pages/dataView/viewPage.jsx";
 
 function AppRoutes() {
   return (
