+++ server/server.js
@@ -2,6 +2,7 @@ import express from "express";
 import cors from "cors";
 import dotenv from "dotenv";
 import cookieParser from "cookie-parser";
+import rateLimit from "express-rate-limit";
 
 import errorHandler from "./middleWare/errorMiddleWare.js";
 import routes from "./routes.js";
@@ -18,10 +19,17 @@ const corsOptions = {
   optionsSuccessStatus: 204,
 };
 
+const limiter = rateLimit({
+  windowMs: 10 * 60 * 1000,
+  max: 100,
+  message: "Too many requests from this IP",
+});
+
 //Middleware
 app.use(cors(corsOptions));
 app.use(express.json());
 app.use(cookieParser());
+app.use(limiter);
 
 //Route middleware
 app.use(errorHandler);
